refactor operation module
